Move UUIDManager interface from utils to framework-db
